requirements.txt
aerich==0.7.2
aiomysql==0.2.0
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgi-lifespan==2.1.0
certifi==2024.8.30
click==8.1.7
dictdiffer==0.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.4
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
iso8601==2.1.0
packaging==24.2
pluggy==1.5.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
PyMySQL==1.1.1
pypika-tortoise==0.2.1
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.41.2
tomlkit==0.13.2
tortoise-orm==0.21.7
typing_extensions==4.12.2
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.0
